requirements.txt
pytest
Selenium
allure-pytest
pyautogui
